backend/requirements.txt
fastapi==0.101.0
uvicorn==0.27.1
python-jose==3.3.0
passlib==1.7.4
python-arango==8.3.1
pydantic-settings==2.13.1